chore(bar): convert stories to typescript
