+++ packages/bar/stories/barCanvas.stories.tsx
@@ -1,7 +1,7 @@
 import { useRef } from 'react'
 import { storiesOf } from '@storybook/react'
 import { generateCountriesData } from '@nivo/generators'
-import { ResponsiveBarCanvas } from '../src'
+import { BarCanvas, BarDatum } from '../src'
 import { button } from '@storybook/addon-knobs'
 
 const keys = ['hot dogs', 'burgers', 'sandwich', 'kebab', 'fries', 'donut']
@@ -9,7 +9,7 @@ const commonProps = {
     width: 900,
     height: 500,
     margin: { top: 60, right: 80, bottom: 60, left: 80 },
-    data: generateCountriesData(keys, { size: 7 }),
+    data: generateCountriesData(keys, { size: 7 }) as BarDatum[],
     indexBy: 'country',
     keys,
     padding: 0.2,
@@ -18,28 +18,24 @@ const commonProps = {
     labelSkipHeight: 16,
 }
 
-const stories = storiesOf('ResponsiveBarCanvas', module)
-
-const Wrapper = props => <div {...props} style={{ height: '300px', width: '600px' }} />
+const stories = storiesOf('BarCanvas', module)
 
 stories.add('custom tooltip', () => (
-    <Wrapper>
-        <ResponsiveBarCanvas
-            {...commonProps}
-            tooltip={({ id, value, color }) => (
-                <strong style={{ color }}>
-                    {id}: {value}
-                </strong>
-            )}
-            theme={{
-                tooltip: {
-                    container: {
-                        background: '#333',
-                    },
+    <BarCanvas
+        {...commonProps}
+        tooltip={({ id, value, color }) => (
+            <strong style={{ color }}>
+                {id}: {value}
+            </strong>
+        )}
+        theme={{
+            tooltip: {
+                container: {
+                    background: '#333',
                 },
-            }}
-        />
-    </Wrapper>
+            },
+        }}
+    />
 ))
 
 stories.add('Get canvas - download the chart', () => {
@@ -53,9 +49,5 @@ stories.add('Get canvas - download the chart', () => {
         link.click()
     })
 
-    return (
-        <Wrapper>
-            <ResponsiveBarCanvas {...commonProps} ref={ref} />
-        </Wrapper>
-    )
+    return <BarCanvas {...commonProps} ref={ref} />
 })
